make group and user method homogeneous
